Обновление ссылок на соцсети
